+++ numpy/cppbuild.sh
@@ -14,6 +14,7 @@ mkdir -p $PLATFORM
 cd $PLATFORM
 INSTALL_PATH=`pwd`
 
+CPYTHON_HOST_PATH="$INSTALL_PATH/../../../cpython/cppbuild/$PLATFORM/host/"
 CPYTHON_PATH="$INSTALL_PATH/../../../cpython/cppbuild/$PLATFORM/"
 OPENBLAS_PATH="$INSTALL_PATH/../../../openblas/cppbuild/$PLATFORM/"
 
@@ -43,13 +44,55 @@ echo "library_dirs = $OPENBLAS_PATH/lib/"         >> site.cfg
 echo "include_dirs = $OPENBLAS_PATH/include/"     >> site.cfg
 
 case $PLATFORM in
-    linux-*)
-        # setup.py won't pick up the right libgfortran.so.4 without this
+    linux-armhf)
+        # setup.py won't pick up the right libgfortran.so without this
+        export LD_LIBRARY_PATH="$OPENBLAS_PATH/lib/:$CPYTHON_PATH/lib/"
+        export PYTHONPATH="$INSTALL_PATH/lib/python3.7/site-packages/"
+        mkdir -p "$PYTHONPATH"
+        "$CPYTHON_HOST_PATH/bin/python3.7" -m pip install --target="$CPYTHON_HOST_PATH/lib/python3.7/" crossenv
+        "$CPYTHON_HOST_PATH/bin/python3.7" -m crossenv "$CPYTHON_PATH/bin/python3.7" crossenv
+        source crossenv/bin/activate
+        ATLAS=None CC="arm-linux-gnueabihf-gcc -std=c99 -march=armv6 -mfpu=vfp -mfloat-abi=hard" python setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ install --prefix $INSTALL_PATH
+        arm-linux-gnueabihf-strip $(find ../ -iname *.so)
+        ;;
+    linux-arm64)
+        # setup.py won't pick up the right libgfortran.so without this
+        export LD_LIBRARY_PATH="$OPENBLAS_PATH/lib/:$CPYTHON_PATH/lib/"
+        export PYTHONPATH="$INSTALL_PATH/lib/python3.7/site-packages/"
+        mkdir -p "$PYTHONPATH"
+        "$CPYTHON_HOST_PATH/bin/python3.7" -m pip install --target="$CPYTHON_HOST_PATH/lib/python3.7/" crossenv
+        "$CPYTHON_HOST_PATH/bin/python3.7" -m crossenv "$CPYTHON_PATH/bin/python3.7" crossenv
+        source crossenv/bin/activate
+        ATLAS=None CC="aarch64-linux-gnu-gcc -mabi=lp64" python setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ install --prefix $INSTALL_PATH
+        aarch64-linux-gnu-strip $(find ../ -iname *.so)
+        ;;
+    linux-ppc64le)
+        # setup.py won't pick up the right libgfortran.so without this
+        export LD_LIBRARY_PATH="$OPENBLAS_PATH/lib/:$CPYTHON_PATH/lib/"
+        export PYTHONPATH="$INSTALL_PATH/lib/python3.7/site-packages/"
+        mkdir -p "$PYTHONPATH"
+        "$CPYTHON_HOST_PATH/bin/python3.7" -m pip install --target="$CPYTHON_HOST_PATH/lib/python3.7/" crossenv
+        "$CPYTHON_HOST_PATH/bin/python3.7" -m crossenv "$CPYTHON_PATH/bin/python3.7" crossenv
+        source crossenv/bin/activate
+        ATLAS=None CC="powerpc64le-linux-gnu-gcc -m64" python setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ install --prefix $INSTALL_PATH
+        powerpc64le-linux-gnu-strip $(find ../ -iname *.so)
+        ;;
+    linux-x86)
+        # setup.py won't pick up the right libgfortran.so without this
         export LD_LIBRARY_PATH="$OPENBLAS_PATH/lib/:$CPYTHON_PATH/lib/"
         export PYTHONPATH="$INSTALL_PATH/lib/python3.7/site-packages/"
         mkdir -p "$PYTHONPATH"
         chmod +x "$CPYTHON_PATH/bin/python3.7"
-        "$CPYTHON_PATH/bin/python3.7" setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ install --prefix $INSTALL_PATH
+        ATLAS=None CC="gcc -m32" "$CPYTHON_PATH/bin/python3.7" setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ install --prefix $INSTALL_PATH
+        strip $(find ../ -iname *.so)
+        ;;
+    linux-x86_64)
+        # setup.py won't pick up the right libgfortran.so without this
+        export LD_LIBRARY_PATH="$OPENBLAS_PATH/lib/:$CPYTHON_PATH/lib/"
+        export PYTHONPATH="$INSTALL_PATH/lib/python3.7/site-packages/"
+        mkdir -p "$PYTHONPATH"
+        chmod +x "$CPYTHON_PATH/bin/python3.7"
+        ATLAS=None CC="gcc -m64" "$CPYTHON_PATH/bin/python3.7" setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ install --prefix $INSTALL_PATH
         strip $(find ../ -iname *.so)
         ;;
     macosx-*)
@@ -57,11 +100,13 @@ case $PLATFORM in
         export PYTHONPATH="$INSTALL_PATH/lib/python3.7/site-packages/"
         mkdir -p "$PYTHONPATH"
         chmod +x "$CPYTHON_PATH/bin/python3.7"
-        "$CPYTHON_PATH/bin/python3.7" setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ install --prefix $INSTALL_PATH
+        ATLAS=None "$CPYTHON_PATH/bin/python3.7" setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ install --prefix $INSTALL_PATH
         # need to add RPATH so it can find MKL in cache
         for f in $(find ../ -iname *.so); do install_name_tool -add_rpath @loader_path/../../../ $f; done
         ;;
-    windows-*)
+    windows-x86)
+        sedinplace '/ccompiler._default_compilers = /,+2d' numpy/distutils/ccompiler.py # don't try to use GCC
+        sedinplace 's/ltype = long_double_representation(pyod("_configtest"))/ltype = "IEEE_DOUBLE_LE"/g' numpy/core/setup_common.py
         CPYTHON_PATH=$(cygpath $CPYTHON_PATH)
         OPENBLAS_PATH=$(cygpath $OPENBLAS_PATH)
         export PATH="$PATH:$OPENBLAS_PATH/:$CPYTHON_PATH/"
@@ -70,7 +115,18 @@ case $PLATFORM in
         # the build sometimes fails with multiple jobs
         MAKEJ=1
         # setup.py install doesn't accept absolute paths on Windows
-        "$CPYTHON_PATH/bin/python.exe" setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ -L$CPYTHON_PATH/libs/ install --prefix ..
+        ATLAS=None "$CPYTHON_PATH/bin/python.exe" setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ -L$CPYTHON_PATH/libs/ install --prefix ..
+        ;;
+    windows-x86_64)
+        CPYTHON_PATH=$(cygpath $CPYTHON_PATH)
+        OPENBLAS_PATH=$(cygpath $OPENBLAS_PATH)
+        export PATH="$PATH:$OPENBLAS_PATH/:$CPYTHON_PATH/"
+        export PYTHONPATH="$INSTALL_PATH/lib/site-packages/"
+        mkdir -p "$PYTHONPATH"
+        # the build sometimes fails with multiple jobs
+        MAKEJ=1
+        # setup.py install doesn't accept absolute paths on Windows
+        ATLAS=None "$CPYTHON_PATH/bin/python.exe" setup.py build -j $MAKEJ build_ext -I$CPYTHON_PATH/include/ -L$CPYTHON_PATH/lib/ -L$CPYTHON_PATH/libs/ install --prefix ..
         ;;
     *)
         echo "Error: Platform \"$PLATFORM\" is not supported"