+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
- * Add `linux-armhf`, `linux-arm64`, and `linux-ppc64le` builds for LLVM ([pull #768](https://github.com/bytedeco/javacpp-presets/pull/768))
+ * Add `linux-armhf`, `linux-arm64`, `linux-ppc64le`, and `windows-x86` builds for CPython, NumPy, and LLVM ([pull #768](https://github.com/bytedeco/javacpp-presets/pull/768))
  * Include `audio_ops.h`, `list_ops.h`, `lookup_ops.h`, and `manip_ops.h` for TensorFlow
  * Add necessary platform properties to build `-gpu` extensions on `linux-arm64` and `linux-ppc64le` ([issue #769](https://github.com/bytedeco/javacpp-presets/issues/769))
  * Add packages missing from TensorFlow ([issue #773](https://github.com/bytedeco/javacpp-presets/issues/773))